Listas de ayudantes, misiones y alumnos

+++ ayudantesDashboard.php
@@ -34,36 +34,52 @@
 
     <!-- contenido -->
 
+
     <div class="container">
         <div class="row">
-            <div class="col s12 l6">
+            <div class="col s12 l8">
                 <?php
-                $sql = "SELECT * FROM ayudante";
-                $result = pg_query_params($dbconn, $sql, array());
-                if (pg_num_rows($result) > 0) {
-                    echo '<table style="width:70%" >';
-                    echo '<tr>';
-                    echo '<th> Rol </th>';
-                    echo '<th> Nombre </th>';
-                    echo '<th> Apellido </th>';
-                    echo '<th> Cantidad de semestres </th>';
-                    while ($row = pg_fetch_assoc($result)) {
+                session_start();
+                $_SESSION["usuario"] =25;
+                $usuario = $_SESSION["usuario"];
+                $sqlprofe = pg_query_params($dbconn,"SELECT idProfesor FROM Profesor WHERE idProfesor = '$usuario' ",array());
+
+                if (pg_num_rows($sqlprofe)!= 0){
+                    $sql = "SELECT * FROM alumno INNER JOIN Ayudantía ON alumno.rolalumno = Ayudantía.rolayudante";
+                    $result = pg_query_params($dbconn, $sql, array());
+                    if (pg_num_rows($result) > 0) {
+                        echo '<table style="width:70%" >';
                         echo '<tr>';
-                        echo '<td>' . $row["rolayudante"] . '</td>';
-                        echo '<td>' . $row["nombre"] . '</td>';
-                        echo '<td>' . $row["apellido"] . '</td>';
-                        echo '<td>' . $row["cantidadsemestres"] . '</td>';
-                        echo '</tr>';
+                        echo '<th> Rol </th>';
+                        echo '<th> Nombre </th>';
+                        echo '<th> Apellido </th>';
+                        echo '<th> Año de ingreso </th>';
+                        echo '<th></th>';
+                        echo '<th></th>';
+                        while ($row = pg_fetch_assoc($result)) {
+                            $rol = $row["rolalumno"];
+                            echo '<tr>';
+                            echo '<td>' . $row["rolalumno"] . '</td>';
+                            echo '<td>' . $row["nombre"] . '</td>';
+                            echo '<td>' . $row["apellido"] . '</td>';
+                            echo '<td>' . $row["añoingreso"] . '</td>';
+                            echo '<td>' . '<a href="modAlumno.php?rolalumno='.$rol.'">modificar</a>' . '</td>';
+                            echo '<td>' . "<a onClick=\"javascript: return confirm('¿Está seguro?');\" href='deleteAyudante.php?idAyudantia=".$row['idayudantia']."'>remover</a>" . '</td>';
+                            echo '</tr>';
+                        }
+                        echo '</table>';
+                        pg_close($dbconn);
+                    } else if (pg_num_rows($result) == 0) {
+                        echo "Aun no hay ayudantes.";
+                        pg_close($dbconn);
+                    } else {
+                        echo "Error al cargar ayudantes";
+                        pg_close($dbconn);
                     }
-                    echo '</table>';
-                    pg_close($dbconn);
-                } else if (pg_num_rows($result) == 0) {
-                    echo "Aun no hay adyudantes.";
-                    pg_close($dbconn);
-                } else {
-                    echo "Error al cargar ayudantes";
-                    pg_close($dbconn);
-                } ?>
+                }else {
+                    echo "<h2>Usuario no autorizado</h2>";
+                }
+                 ?>
 
             </div>
             <div class="col s12 l6">
